requirements.txt
celery==5.0.5
Django==3.0.2
djangorestframework== 3.12.2
redis==3.5.3
